Fix MemoryPage size

+++ src/Basic/MemoryUtils/Models/MemoryPage.cs
@@ -5,14 +5,14 @@ public record MemoryPage
     internal unsafe MemoryPage(MEMORY_BASIC_INFORMATION mbi)
     {
         Base = (nint)mbi.BaseAddress;
-        RegionSize = (int)mbi.RegionSize;
+        RegionSize = (nint)mbi.RegionSize;
         Protect = mbi.Protect;
         State = mbi.State;
         Type = mbi.Type;
     }
 
     public nint Base { get; }
-    public int RegionSize { get; }
+    public nint RegionSize { get; }
 
     public MemPageProtect Protect { get; }
     public MemPageState State { get; }
